+++ IBAnimatableApp/BorderViewController.swift
@@ -13,7 +13,10 @@ class BorderViewController: UIViewController {
   @IBOutlet weak var botCheckBox: AnimatableCheckBox!
   @IBOutlet weak var leftCheckBox: AnimatableCheckBox!
   @IBOutlet weak var rightCheckBox: AnimatableCheckBox!
-  
+  @IBOutlet weak var topOutCheckBox: AnimatableCheckBox!
+  @IBOutlet weak var botOutCheckBox: AnimatableCheckBox!
+  @IBOutlet weak var leftOutCheckBoxOut: AnimatableCheckBox!
+  @IBOutlet weak var rightOutCheckBox: AnimatableCheckBox!
 
   @IBAction func boxChecked(_ sender: AnimatableCheckBox) {
     let border: BorderSides
@@ -26,9 +29,18 @@ class BorderViewController: UIViewController {
       border = .left
     case rightCheckBox:
       border = .right
+    case topOutCheckBox:
+        border = .topOut
+    case botOutCheckBox:
+        border = .bottomOut
+    case leftOutCheckBoxOut:
+        border = .leftOut
+    case rightOutCheckBox:
+        border = .rightOut
     default:
       return
     }
+
     if sender.checked {
       viewToBorder.borderSides.insert(border)
     } else {